+++ static/js/log.js
@@ -5,6 +5,16 @@ $(function () {
 
         $.get(this.dataset.url, function(data) {
             $("#ping-details-body").html(data);
+
+            var htmlPre = $("#email-body-html pre");
+            if (htmlPre.length) {
+                var opts = {USE_PROFILES: {html: true}};
+                var clean = DOMPurify.sanitize(htmlPre.text(), opts);
+                var blob = new Blob([clean], {type: "text/html"});
+
+                htmlPre.remove();
+                document.getElementById("email-body-html-iframe").src = URL.createObjectURL(blob);
+            }
         });
 
         return false;
@@ -16,8 +26,8 @@ $(function () {
             format == "local" ? dt.local() : dt.tz(format);
 
             $(".date", row).text(dt.format("MMM D"));
-            $(".time", row).text(dt.format("HH:mm"));                
-        })        
+            $(".time", row).text(dt.format("HH:mm"));
+        })
     }
 
     $("#format-switcher").click(function(ev) {